requirements.txt
pytest==8.0.1
autopep8==2.0.4
python-dotenv==1.0.1
